Fix scheme to work with forwarded requests
Backported from later rack release.

This fixes request.url to use the actual protocol the client used.
As we run behind AWS ELB that terminates SSL and forwards the request with http to our app we need to take the HTTP_X_FORWARDED_PROTO header instead of the protocol used by ELB.

Needed this as our Meter middleware extracts the url from Rack and would log the wrong url.